+++ src/library/audio/sdl/sdlaudio.cpp
@@ -41,7 +41,9 @@ static const char* dummySDLDriver = "libtas";
 static SDL_AudioFormat audioFormat;
 // static Uint16 bufferSamplesSize;
 
-static std::shared_ptr<AudioSource> sourceSDL;
+/* All SDL devices. From SDL implementation, there is a limit of 16 devices */
+#define MAX_SDL_SOURCES 16
+static std::shared_ptr<AudioSource> sourcesSDL[MAX_SDL_SOURCES];
 
 static const char* dummySDLDevice = "libTAS device";
 static std::string curDriver;
@@ -101,12 +103,26 @@ char * SDL_AudioDriverName(char *namebuf, int maxlen)
     return nullptr;
 }
 
-/* Override */ int SDL_OpenAudio(SDL_AudioSpec * desired, SDL_AudioSpec * obtained)
+/* Helper function for SDL_OpenAudio() and SDL_OpenAudioDevice() */
+static int open_audio_device(const SDL_AudioSpec * desired, SDL_AudioSpec * obtained, int min_id)
 {
-    DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-
     if (shared_config.audio_disabled)
         return -1;
+
+    /* Look for the next available device. Id 1 is reserved for SDL_OpenAudio() device */
+    int id = min_id;
+    for (; id < MAX_SDL_SOURCES; id++)
+        if (!sourcesSDL[id-1])
+            break;
+
+    if (sourcesSDL[id-1]) {
+        debuglogstdio(LCF_SDL | LCF_SOUND, "   no available device");
+        return -1;
+    }
+
+    if (obtained != NULL) {
+        memmove(obtained, desired, sizeof(SDL_AudioSpec));
+    }
 
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
 
@@ -114,14 +130,14 @@ char * SDL_AudioDriverName(char *namebuf, int maxlen)
     auto buffer = audiocontext.getBuffer(bufferId);
 
     /* Sanity check done by SDL */
-    if (!desired->freq) desired->freq = 22050;
-    buffer->frequency = desired->freq;
+    if (!obtained->freq) obtained->freq = 22050;
+    buffer->frequency = obtained->freq;
     debuglogstdio(LCF_SDL | LCF_SOUND, "Frequency %d Hz", buffer->frequency);
 
     /* Sanity check done by SDL */
-    if (!desired->format) desired->format = AUDIO_S16LSB;
+    if (!obtained->format) obtained->format = AUDIO_S16LSB;
 
-    switch(desired->format) {
+    switch(obtained->format) {
         case AUDIO_U8:
             buffer->format = AudioBuffer::SAMPLE_FMT_U8;
             break;
@@ -139,24 +155,24 @@ char * SDL_AudioDriverName(char *namebuf, int maxlen)
             return -1;
     }
     /* Sanity check done by SDL */
-    if (!desired->channels) desired->channels = 2;
-    buffer->nbChannels = desired->channels;
+    if (!obtained->channels) obtained->channels = 2;
+    buffer->nbChannels = obtained->channels;
     debuglogstdio(LCF_SDL | LCF_SOUND, "Channels %d", buffer->nbChannels);
 
     buffer->update();
     debuglogstdio(LCF_SDL | LCF_SOUND, "Format %d bits", buffer->bitDepth);
 
-    buffer->size = desired->samples * buffer->alignSize;
+    buffer->size = obtained->samples * buffer->alignSize;
     buffer->update(); // Yes, a second time, to fill sampleSize based on size.
     buffer->samples.resize(buffer->size);
 
     /* Push buffers in a source */
     int sourceId = audiocontext.createSource();
-    sourceSDL = audiocontext.getSource(sourceId);
+    sourcesSDL[id-1] = audiocontext.getSource(sourceId);
 
-    sourceSDL->buffer_queue.push_back(buffer);
+    sourcesSDL[id-1]->buffer_queue.push_back(buffer);
 
-    if (desired->callback != NULL) {
+    if (obtained->callback != NULL) {
         /* We are using the callback mechanism.
          * When the buffer is depleted, we are supposed to call the
          * callback function so that it fills the buffer again.
@@ -165,10 +181,10 @@ char * SDL_AudioDriverName(char *namebuf, int maxlen)
          * it calls the callback function to fill the buffer again.
          */
 
-        sourceSDL->source = AudioSource::SOURCE_CALLBACK;
-        SDL_AudioCallback audioCallback = desired->callback;
-        void* callbackArg = desired->userdata;
-        sourceSDL->callback = [audioCallback, callbackArg](AudioBuffer& ab){
+        sourcesSDL[id-1]->source = AudioSource::SOURCE_CALLBACK;
+        SDL_AudioCallback audioCallback = obtained->callback;
+        void* callbackArg = obtained->userdata;
+        sourcesSDL[id-1]->callback = [audioCallback, callbackArg](AudioBuffer& ab){
             /* Emptying the audio buffer */
             ab.makeSilent();
 
@@ -185,25 +201,50 @@ char * SDL_AudioDriverName(char *namebuf, int maxlen)
         /* We are using the push mechanism. The game will use
          * SDL_QueueAudio() to push audio buffers.
          */
-        sourceSDL->source = AudioSource::SOURCE_STREAMING_CONTINUOUS;
+        sourcesSDL[id-1]->source = AudioSource::SOURCE_STREAMING_CONTINUOUS;
     }
     /* We simulate an empty buffer by setting the position at the end */
-    sourceSDL->position = buffer->sampleSize;
+    sourcesSDL[id-1]->position = buffer->sampleSize;
 
     /* We must fill some information in the structs */
-    desired->size = buffer->size;
+    obtained->size = buffer->size;
 
     /* Filling silence value. Not sure what to put here */
-    desired->silence = (buffer->format==AudioBuffer::SAMPLE_FMT_U8)?0x80:0x00;
-
-    if (obtained != NULL) {
-        memmove(obtained, desired, sizeof(SDL_AudioSpec));
-    }
+    obtained->silence = (buffer->format==AudioBuffer::SAMPLE_FMT_U8)?0x80:0x00;
 
     /* Keep a copy of the audio format */
-    audioFormat = desired->format;
+    audioFormat = obtained->format;
 
-    return 0;
+    return id;
+}
+
+/* Override */ int SDL_OpenAudio(SDL_AudioSpec * desired, SDL_AudioSpec * obtained)
+{
+    DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
+
+    /* SDL1 has only one device which is id 1 */    
+    if (sourcesSDL[0]) {
+        debuglogstdio(LCF_SDL | LCF_SOUND, "   audio device already opened");
+        return -1;
+    }
+
+    SDL_AudioDeviceID id = 0;
+    
+    if (obtained) {
+        id = open_audio_device(desired, obtained, 1);
+    }
+    else {
+        SDL_AudioSpec _obtained;
+        memset(&_obtained, 0, sizeof(SDL_AudioSpec));
+        id = open_audio_device(desired, &_obtained, 1);
+        /* On successful open, copy calculated values into 'desired'. */
+        if (id > 0) {
+            desired->size = _obtained.size;
+            desired->silence = _obtained.silence;
+        }
+    }
+    
+    return id==1?0:-1;
 }
 
 /* Override */ int SDL_GetNumAudioDevices(int iscapture)
@@ -228,18 +269,27 @@ char * SDL_AudioDriverName(char *namebuf, int maxlen)
                    int iscapture, const SDL_AudioSpec *desired,
                    SDL_AudioSpec *obtained, int allowed_changes)
 {
-    DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
+    debuglogstdio(LCF_SDL | LCF_SOUND, "%s called for device %s", __func__, device?device:"NULL");
     if (iscapture != 0)
         return 0;
 
-    int ret = SDL_OpenAudio(const_cast<SDL_AudioSpec*>(desired), obtained); // TODO: cast probably not good
-    return (ret==-1)?0:2;
+    return open_audio_device(desired, obtained, 2);
 }
 
 /* Override */ SDL_AudioStatus SDL_GetAudioStatus(void)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-    switch(sourceSDL->state) {
+    return SDL_GetAudioDeviceStatus(1);
+}
+
+/* Override */ SDL_AudioStatus SDL_GetAudioDeviceStatus(SDL_AudioDeviceID dev)
+{
+    DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
+
+    if (!sourcesSDL[dev-1])
+        return SDL_AUDIO_STOPPED;
+
+    switch(sourcesSDL[dev-1]->state) {
         case AudioSource::SOURCE_INITIAL:
         case AudioSource::SOURCE_STOPPED:
             return SDL_AUDIO_STOPPED;
@@ -253,29 +303,23 @@ char * SDL_AudioDriverName(char *namebuf, int maxlen)
     }
 }
 
-/* Override */ SDL_AudioStatus SDL_GetAudioDeviceStatus(SDL_AudioDeviceID dev)
-{
-    DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-    return SDL_GetAudioStatus();
-}
-
 /* Override */ void SDL_PauseAudio(int pause_on)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-    std::lock_guard<std::mutex> lock(audiocontext.mutex);
-    if (!sourceSDL)
-        return;
-    if (pause_on == 0)
-        sourceSDL->state = AudioSource::SOURCE_PLAYING;
-    else
-        sourceSDL->state = AudioSource::SOURCE_PAUSED;
-
+    SDL_PauseAudioDevice(pause_on, 1);
 }
 
 /* Override */ void SDL_PauseAudioDevice(SDL_AudioDeviceID dev, int pause_on)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-    SDL_PauseAudio(pause_on);
+    if (!sourcesSDL[dev-1])
+        return;
+
+    std::lock_guard<std::mutex> lock(audiocontext.mutex);
+    if (pause_on == 0)
+        sourcesSDL[dev-1]->state = AudioSource::SOURCE_PLAYING;
+    else
+        sourcesSDL[dev-1]->state = AudioSource::SOURCE_PAUSED;
 }
 
 void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
@@ -303,7 +347,10 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
 {
     debuglogstdio(LCF_SDL | LCF_SOUND, "%s call with %d bytes of data", __func__, len);
 
-    if (sourceSDL->source == AudioSource::SOURCE_CALLBACK) {
+    if (!sourcesSDL[dev-1])
+        return -1;
+
+    if (sourcesSDL[dev-1]->source == AudioSource::SOURCE_CALLBACK) {
         /* We cannot queue samples when using the callback mechanism */
         return -1;
     }
@@ -312,11 +359,11 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
 
     /* We try to reuse a buffer that has been processed from the source */
     std::shared_ptr<AudioBuffer> ab;
-    if (sourceSDL->nbQueueProcessed() > 0) {
+    if (sourcesSDL[dev-1]->nbQueueProcessed() > 0) {
         /* Removing first buffer */
-        ab = sourceSDL->buffer_queue[0];
-        sourceSDL->buffer_queue.erase(sourceSDL->buffer_queue.begin());
-        sourceSDL->queue_index--;
+        ab = sourcesSDL[dev-1]->buffer_queue[0];
+        sourcesSDL[dev-1]->buffer_queue.erase(sourcesSDL[dev-1]->buffer_queue.begin());
+        sourcesSDL[dev-1]->queue_index--;
     }
     else {
         /* Building a new buffer */
@@ -324,12 +371,12 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
         ab = audiocontext.getBuffer(bufferId);
 
         /* Getting the parameters of the buffer from one of the queue */
-        if (sourceSDL->buffer_queue.empty()) {
+        if (sourcesSDL[dev-1]->buffer_queue.empty()) {
             debuglogstdio(LCF_SDL | LCF_SOUND | LCF_ERROR, "Empty queue, cannot guess buffer parameters");
             return -1;
         }
 
-        auto ref = sourceSDL->buffer_queue[0];
+        auto ref = sourcesSDL[dev-1]->buffer_queue[0];
         ab->format = ref->format;
         ab->nbChannels = ref->nbChannels;
         ab->frequency = ref->frequency;
@@ -340,7 +387,7 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
     ab->samples.insert(ab->samples.end(), static_cast<const uint8_t*>(data), &(static_cast<const uint8_t*>(data))[len]);
     ab->size = len;
     ab->update();
-    sourceSDL->buffer_queue.push_back(ab);
+    sourcesSDL[dev-1]->buffer_queue.push_back(ab);
 
     return 0;
 }
@@ -349,13 +396,16 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
 
-    if (sourceSDL->source == AudioSource::SOURCE_CALLBACK) {
+    if (!sourcesSDL[dev-1])
+        return -1;
+
+    if (sourcesSDL[dev-1]->source == AudioSource::SOURCE_CALLBACK) {
         /* We cannot get queue samples when using the callback mechanism */
         return 0;
     }
 
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
-    Uint32 qsize = sourceSDL->queueSize() - sourceSDL->getPosition();
+    Uint32 qsize = sourcesSDL[dev-1]->queueSize() - sourcesSDL[dev-1]->getPosition();
     debuglogstdio(LCF_SDL | LCF_SOUND, "Returning %d", qsize);
     return qsize;
 }
@@ -364,7 +414,10 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
 
-    if (sourceSDL->source == AudioSource::SOURCE_CALLBACK) {
+    if (!sourcesSDL[dev-1])
+        return;
+
+    if (sourcesSDL[dev-1]->source == AudioSource::SOURCE_CALLBACK) {
         /* We cannot get queue samples when using the callback mechanism */
         return;
     }
@@ -373,15 +426,13 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
     /* We simulate clearing the queue by setting the position to the end
      * of the queue.
      */
-    sourceSDL->setPosition(sourceSDL->queueSize());
+    sourcesSDL[dev-1]->setPosition(sourcesSDL[dev-1]->queueSize());
 }
 
 /* Override */ void SDL_LockAudio(void)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-    if (ThreadManager::getThreadId() == audiocontext.audio_thread)
-        return;
-    mutex.lock();
+    return SDL_LockAudioDevice(1);
 }
 
 /* Override */ void SDL_LockAudioDevice(SDL_AudioDeviceID dev)
@@ -395,9 +446,7 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
 /* Override */ void SDL_UnlockAudio(void)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-    if (ThreadManager::getThreadId() == audiocontext.audio_thread)
-        return;
-    mutex.unlock();
+    return SDL_UnlockAudioDevice(1);
 }
 
 /* Override */ void SDL_UnlockAudioDevice(SDL_AudioDeviceID dev)
@@ -411,15 +460,7 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
 /* Override */ void SDL_CloseAudio(void)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
-
-    std::lock_guard<std::mutex> lock(audiocontext.mutex);
-    /* Remove the source from the audio context */
-    if (sourceSDL) {
-        audiocontext.deleteSource(sourceSDL->id);
-    }
-
-    /* Destroy the source object */
-    sourceSDL.reset();
+    return SDL_CloseAudioDevice(1);
 }
 
 /* Override */ void SDL_CloseAudioDevice(SDL_AudioDeviceID dev)
@@ -428,12 +469,12 @@ void SDL_MixAudio(Uint8 * dst, const Uint8 * src, Uint32 len, int volume)
 
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     /* Remove the source from the audio context */
-    if (sourceSDL) {
-        audiocontext.deleteSource(sourceSDL->id);
+    if (sourcesSDL[dev-1]) {
+        audiocontext.deleteSource(sourcesSDL[dev-1]->id);
     }
 
     /* Destroy the source object */
-    sourceSDL.reset();
+    sourcesSDL[dev-1].reset();
 }
 
 }