+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add lua scripting
 * Check for gdb presence
 * Use vdpau-va-gl for vdpau software rendering (#390)
+* Implement multiple SDL audio devices
 
 ### Changed